+++ tests/test_local_backend.py
@@ -40,5 +40,9 @@ class LocalBackendTest(BackendTestCase):
         root = self.test_dir.join('default')
         backend = LocalBackend('default', Config({}))
         assert backend.root == root
-        # with app.app_context():
-        #     self.assertEqual(backend.root, root)
+
+    def test_backend_root(self, app):
+        app.config['LOCAL_FS_ROOT'] = str(self.test_dir)
+        root = self.test_dir.join('default')
+        backend = LocalBackend('default', Config({}))
+        assert backend.root == root