+++ CHANGELOG.rst
@@ -4,7 +4,7 @@ Changelog
 Current
 -------
 
-- Added backend level configuration
+- Added backend level configuration ``FS_{BACKEND_NAME}_{KEY}``
 
 0.3.0 (2017-03-05)
 ------------------